I implemented the time-threshold solution for QWI as we discussed last week.  To be specific:
1. I disabled the probability-threshold from last week.
2. I wait until the probability at the screen reaches 10^-4 (numerical artifacts reach the screen long before the visible wave does), then start counting.
3. 30 time steps after starting the count, I allow detections to happen.

This should solve the problem we wanted to solve with the probability-threshold, namely, the intensity pattern for washes out if detections happen too soon.  Also, this should hopefully solve some associated problems with the probability-threshold implementation, for example blocking one of the double slits reduces the magnutide of the transmitted wave so much that nothing is ever detected on the screen.

One worry I have about this is the fact that it causes almost all (99% or more?) detections to happen at about the same time; maybe this will give the wrong idea to students?

Let me know how it goes, including:
1. Whether you'd like me to increase/decrease the 30 steps (or add a developer control)
2. Should this replace the main site version?