PhetPCanvas.updateLayout was previously called for any PhetPCanvas whose size as changed, even if the canvas was associated with an inactive module.  This causes a performance problem when resizing, since all canvases will have their layouts updated.
This fix calls updateLayout only if the PhetPCanvas is showing on the screen.
If it's not showing, a layoutDirty flag is set, and the call to
updateLayout is deferred until when the canvas or one of its ancestors
is made visible.

Subclasses the override updateLayout must now call super.updateLayout
to properly clear the layoutDirty flag.