website: Speed improvements (many changes)
This is related to the previous batch of speed changes, I had found more calls that automatically format the data as it comes out of the database.  I've gone through and change the assumption that data from the database is always formatted.  Instead, it should only be formatted if it is rendered.