website: Refactored page templates & cache enhancement
Took all the PhET website specific stuff out of BasePage and placed it in SitePage where it belongs.  This includes caching.

SitePage now does more with caching, and the update will not happen if the page should be served from the cache instead.

Unfortunately, the way pages deal with SitePage is inconsistent and needs some attention.  Add this to the TODO list.