website: Simple but prolific changes: minimal code executing in a session, "relative" include paths, NavBar class, better separation of BasePage and SitePage, update and homogonization of pages derived from SitePage, some code cleanup, and fixing of the subscription page, removed some pages from being cached, and added some more cache clearing code when things happen
Session Changes:
We used to just start a session and have it active the whole time the script executing.  However, our web server is having problems with the session grabbing an exclusive lock, and several different scripts (50+) waiting on that lock, but never recover.  Not reproducable, not sure why it happens, so we're trying this hoping it will help with that problem.  It will help in any event if a user has multiple windows into our site.

There are some awkward things done to the cache to accept this change.  Not to pretty but beats a rewrite for now.

To be able to do this, accessing cookies (currently only done in the authentication code) needs to happen close together, temporally.  So all references directly to the authentication code have been minimized (there are still a few), and new accessor functions have been created in SitePage where appropriate.

"Relative" include paths
ALL pages define SITE_ROOT, the relative path pointing to the root of the website (usually "../" but sometimes is different).  This help get around PHP's weird behavior where files included from another include file (chaining) is _always_ relative to the original _calling_ file, rather than the included file.

There is also a PORTAL_ROOT that points to one level higher.

Relative paths should use these and only hardcode when necessary.


NavBar Class & Increased BasePage and Site Page separation
Took the navigation bar stuff and put it in a class that is a member of site page.  BasePage has hooks to call the navigation bar, but these can be ignored or overidden as necessary.  Other more minor enhancements between SitePage and BasePage.

The above 2 things finally allows the root index.php page to derive from SitePage, rather than having to do everything on its own.  The new root index.php can use some cleanup, but at least it is more consistent with everything else.


Homogonize & Update pages derived from SitePage
Since I was in there, I went through and updated most pages for consistancy of handling.  Sometimes this is just whitespace changes, sometimes this is more.


Subscription Page
It was broken.  Something I did?  Not sure, but it was needlessly complex anyway so I refactored it.  Now it is simple and it works properly.