website: Newsletter now working
This involves using the Mail_Queue PEAR library, which also needs the MDB2#mysql and Net_SMTP libraries.  The newsletter happens in 2 steps: first it is composed and each email is queued for distribution.  The a cron script will do the actual delivery late at night when the server is less loaded.