website: Resolving an issue in IE with the multi-select element and the mousewheel
Old behavior: select something with the multi-select element and it will be added to the list.  Use the mousewheel and the select box part of the multi-select element will scroll through the options, adding each to the list.

New behavior: mousewheel has no effect.  This has a side effect, where if the select box list is open and it has a scroll bar, it will no scroll as one would expect.  This is more desireable than the bug, and so it will stay.