website: Fixed Excel pages downloading with wrong mime type
This is much like the PPT fix earlier.  'file' will identify office documens as MS Word files, even if they are Powerpoint or Exel.   Suprise suprise, this causes problems for IE, but not other browsers.  The fix involves checking the three letter extension.  Easy to circumvent, we'll hope that our contributors are nice.