updates from John Henson:
I found an error in the autoTrack script. This is an error on MY part, not yours.  This was a line of code that didn’t get copied when I created the phet-version of the script. (I’m using a similar script across organizations and sites).

If you could use the modified version instead, we should be able to start tracking downloads.