rename OutputStreamReader to StreamReaderThread.
This class reads from an input stream, not an output stream, and does so in a separate thread.  Could also have renamed this to InputStreamReader, but having both Input and Reader in the name is redundant, since you can only read from an input.  But having Thread in the name is useful, since does the reading in a separate thread.