website: Early rewrite of update-localized-jars.php in python, adding pre-caching of localized downloadable Flash JAR files.  Still needs error handling.
CL PHP is combersome to develop as just about any different environment in which it is run presents a new set of challenges.  Python is quicker, easier, and funner (for me), so we'll go with that.  It works bit I'm not confident it isn't ready others to play with yet.