The branch hierarchical-categories is to refine the website feature of (suprise suprise) having hierarchical categories in the NavBar menus.
The requested behavior is to snap shut.  They don't currently snap
shut.

At the time they were first introduced, there were a couple technical
issues that prevented this in the initial implementation.  Since then
a couple code roadblocks have been removed, and combined with a
couple of creative ideas has raised the possibility that they might be
possible without much effort.

Since the behavior will need approval before being incorporated into
the main site it will live in this branch.  It will only be deployed
to phet-server when it is ready for feedback.  When given Caesar's
thumb it will rejoin to the main branch or die out here on the
frontier.