The NavBar now acts more like a 'history' rather than an 'address' for simulation pages.
Before: if you click on a simulation from the teacher ideas browse page (or anywhere else), the NavBar would open the top-level "Simulations" section of the navigation menu.

Now: it will stay in whatever section you navigated to.  So if you were in the teacher ideas browse page, the NavBar will keep that section open.

This is only for the simulation pages.